chore(domain): move task status and task failure reason classes in separate module
