+++ internal/domain/task/value_objects/task_status.py
@@ -1,0 +1,8 @@
+from enum import StrEnum, auto
+
+
+class TaskStatus(StrEnum):
+    FAILED = auto()
+    CREATED = auto()
+    RUNNING = auto()
+    COMPLETED = auto()
